Add `constraints` schema (#2)
This Pull Request introduces a feature that allows users to include an additional validation function when adding a value to the option beyond just validating the value type. The function may, for example, validate whether a particular value conforms to a specified pattern. Users can now integrate this validation function by including it in the `constraints` within the schema.

```php
$option = new Option($this->hook, null, 1);
$option->setSchema([
   'foo' => [
      'type' => 'string',
      'constraints' => 'is_email',
   ],
]);
$option->register();

add_option('foo', 'not email'); // Will throw an Exception.
```

This enables users to enhance validation for specific values, giving greater flexibility and control.